backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
psycopg2-binary==2.9.10
gunicorn==20.1.0
